requirements.txt
requests
beautifulsoup4
python-dateutil
markdownify
python-dotenv
lxml
